Enables path comparison by regexp
* Passes comparison result to corresponding handler via request object.

+++ src/lib/http/http_handle_connection.c
@@ -3,7 +3,7 @@
 int http_handle_connection(RequestContext *context, Client client) {
   Router *router = (Router*)context->router;
   puts("http/handle_connection.c");
-  for(int i = 0; i < ARRAY_LEN(router->patterns) && router->patterns[i]; i++) {
+  for(size_t i = 0; i < ARRAY_LEN(router->patterns) && router->patterns[i]; i++) {
     printf("HCRT: %s\n", router->patterns[i]);
   }
   printf("PROCESSOR COMPONENTS: %s\n", router->patterns[1]);
